+++ Interface/AddOns/MissingTradeSkillsList/UI/Common/FilterFrame.lua
@@ -12,12 +12,9 @@ MTSLUI_FILTER_FRAME = {
     FRAME_WIDTH_HORIZONTAL = 515,
     -- height of the frame
     FRAME_HEIGHT = 110,
-    -- keep track of current sort mehod (1 = name (default), 2 = level)
-    current_sort = 1,
     -- keeps track of current phase used for filtering
     current_phase,
     phases= {},
-    sorts = {},
     -- all contintents
     continents = {},
     -- all zones for each contintent
@@ -52,14 +49,6 @@ MTSLUI_FILTER_FRAME = {
         self:InitialiseData()
         -- create the container frame
         self.ui_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", parent_frame, nil, self.FRAME_WIDTH_VERTICAL, self.FRAME_HEIGHT, false)
-        -- create a filter for sorting
-       -- self.ui_frame.sort_by_text = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["sort"][MTSLUI_CURRENT_LANGUAGE], 5, -4, "LABEL", "TOPLEFT")
-       -- self.ui_frame.sort_drop_down = CreateFrame("Frame", filter_frame_name .. "_DD_SORTS", self.ui_frame, "UIDropDownMenuTemplate")
-       -- self.ui_frame.sort_drop_down:SetPoint("TOPLEFT", self.ui_frame.sort_by_text, "TOPRIGHT", -10, 9)
-       -- self.ui_frame.sort_drop_down.filter_frame_name = filter_frame_name
-       -- self.ui_frame.sort_drop_down.initialize = self.CreateDropDownSorting
-       -- UIDropDownMenu_SetWidth(self.ui_frame.sort_drop_down, 70)
-       -- UIDropDownMenu_SetText(self.ui_frame.sort_drop_down, self.sorts[self.current_sort]["name"])
         -- Search box with button
         self.ui_frame.search_box = CreateFrame("EditBox", filter_frame_name .. "_TF", self.ui_frame)
         self.ui_frame.search_box:SetBackdrop({
@@ -136,6 +125,28 @@ MTSLUI_FILTER_FRAME = {
         _G[filter_frame_name] = self
     end,
 
+    ----------------------------------------------------------------------------------------------------------
+    -- Reset the filters to default value (when parent window is hidden/closed)
+    ----------------------------------------------------------------------------------------------------------
+    ResetFilters = function(self)
+        -- reset name to search
+        self.ui_frame.search_box:SetText("")
+        self.ui_frame.search_box:ClearFocus()
+        -- reset source type
+        self.current_source_id = "any"
+        UIDropDownMenu_SetText(self.ui_frame.source_drop_down, MTSLUI_LOCALES_LABELS["any"][MTSLUI_CURRENT_LANGUAGE])
+        -- reset phase to current
+        self.current_phase = 1
+        UIDropDownMenu_SetText(self.ui_frame.phase_drop_down, MTSLUI_LOCALES_LABELS["current"][MTSLUI_CURRENT_LANGUAGE] .. " (" .. MTSL_CURRENT_PHASE .. ")")
+        -- reset specialization
+        self.current_spec_id = 0
+        UIDropDownMenu_SetText(self.ui_frame.specs_drop_down, MTSLUI_LOCALES_LABELS["any"][MTSLUI_CURRENT_LANGUAGE])
+        -- reset contintent & zone
+        self.current_contintent_id = 0
+        UIDropDownMenu_SetText(self.ui_frame.continent_drop_down, MTSLUI_LOCALES_LABELS["any"][MTSLUI_CURRENT_LANGUAGE])
+        UIDropDownMenu_SetText(self.ui_frame.zone_drop_down, "")
+    end,
+
     -- Limit the filter for phase to current only
     UseOnlyCurrentPhase = function(self)
         self.phases = {
@@ -157,26 +168,11 @@ MTSLUI_FILTER_FRAME = {
     -- Build the fixed arrays with all continents & zones available
     ----------------------------------------------------------------------------------------------------------
     InitialiseData = function(self)
-        -- self:BuildSorts()
         self:BuildPhases()
         self:BuildSources()
         self:BuildSpecializations()
         self:BuildContinents()
         self:BuildZones()
-    end,
-
-    BuildSorts = function(self)
-        self.sorts = {
-            {
-                ["name"] = MTSLUI_LOCALES_LABELS["level"][MTSLUI_CURRENT_LANGUAGE],
-                ["id"] = 1,
-            },
-            {
-                ["name"] = MTSLUI_LOCALES_LABELS["name"][MTSLUI_CURRENT_LANGUAGE],
-                ["id"] = 2,
-            },
-        }
-        self.current_sort = 1
     end,
 
     BuildPhases = function(self)
@@ -423,32 +419,6 @@ MTSLUI_FILTER_FRAME = {
         -- Apply filter if we may
         if self:IsFilteringEnabled() then
             self.list_frame:ChangeZone(id)
-        end
-    end,
-
-    ----------------------------------------------------------------------------------------------------------
-    -- Intialises drop down for sorting
-    ----------------------------------------------------------------------------------------------------------
-    CreateDropDownSorting = function(self)
-        MTSLUI_TOOLS:FillDropDown( _G[self.filter_frame_name].sorts, _G[self.filter_frame_name].ChangeSortHandler, self.filter_frame_name)
-    end,
-
-    --------------------------------------------------------------------------------------
-    -- Handles DropDown Change event after changing the sorting
-    ----------------------------------------------------------------------------------------------------------
-    ChangeSortHandler = function(self, value, text)
-        -- Only trigger update if we change to a different way to sort
-        if value ~= nil and self.current_sort ~= value then
-            self:ChangeSorting(value, text)
-        end
-    end,
-
-    ChangeSorting = function(self, value, text)
-        self.current_sort = value
-        UIDropDownMenu_SetText(self.ui_frame.sort_drop_down, text)
-        -- Apply filter if we may
-        if self:IsFilteringEnabled() then
-            self.list_frame:ChangeSort(value)
         end
     end,