+++ Interface/AddOns/MissingTradeSkillsList/UI/Common/SkillDetailFrame.lua
@@ -63,11 +63,11 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.ui_frame:SetBackdropColor(0,0,0,1)
         -- Add the Texts/Strings to the frame
         local text_label_left = 10
-        local text_label_right = 125
+        local text_label_right = 130
         local text_label_top = -8
         local text_gap = 16
         -- Labels to show "name: <name>"
-        self.labels.name.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["name"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.name.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["name"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.name.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         -- create a frame to hover over and show the tooltip
         self.labels.name.tooltip_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", self.ui_frame, nil, 300, text_gap + 5, false)
@@ -76,31 +76,31 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.name.tooltip_frame:SetScript("OnLeave", function() _G.GameTooltip:Hide() end)
         text_label_top = text_label_top - text_gap
         -- Labels to show "Required skill: <min skill>"
-        self.labels.min_skill.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs skill level"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.min_skill.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs skill level"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.min_skill.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Required XP level: <xp level>"
-        self.labels.requires_xp.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs XP level"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.requires_xp.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs XP level"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.requires_xp.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Required reputation: <reputation>"
-        self.labels.requires_rep.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs reputation"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.requires_rep.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs reputation"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.requires_rep.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Required reputation: <reputation>"
-        self.labels.requires_spec.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs specialization"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.requires_spec.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["needs specialization"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.requires_spec.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Special action <special action>"
-        self.labels.special_action.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["special action"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.special_action.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["special action"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.special_action.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Price: <price>"
-        self.labels.price.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["price"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.price.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["price"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.price.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         text_label_top = text_label_top - text_gap
         -- Labels to show "Learned from: <type>"
-        self.labels.type.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.type.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.type.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         -- create a frame to hover over and show the tooltip
         self.labels.type.tooltip_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", self.ui_frame, nil, 300, text_gap + 5, false)
@@ -109,7 +109,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.type.tooltip_frame:SetScript("OnLeave", function() _G.GameTooltip:Hide() end)
         text_label_top = text_label_top - text_gap
         -- Labels to show "Trained by: <trainers> or Sold by: <vendors> or Dropped by: <mobs> or Obtained from: <quest>"
-        self.labels.sources.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.sources.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.sources.title:Hide()
         -- Create a frame to hold the labels
         self.labels.sources.ui_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", self.ui_frame, nil, self.FRAME_WIDTH - 107, self.FRAME_SOURCES_HEIGHT, false)
@@ -129,9 +129,9 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.sources.ui_frame:SetScript("OnMouseUp", function() event_class:AddTomTomWayPointPrimarySource() end)
         -- hide on creation
         self.labels.sources.ui_frame:Show()
-        text_label_right = 125
+        text_label_right = 130
         text_label_top = text_label_top + text_gap + 2
-        self.labels.alt_type.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["also learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.alt_type.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["also learned from"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         self.labels.alt_type.value = MTSLUI_TOOLS:CreateLabel(self.ui_frame, "-", text_label_right, text_label_top, "TEXT", "TOPLEFT")
         -- create a frame to hover over and show the tooltip
         self.labels.alt_type.tooltip_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", self.ui_frame, nil, 300, text_gap + 5, false)
@@ -140,7 +140,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.alt_type.tooltip_frame:SetScript("OnLeave", function() _G.GameTooltip:Hide() end)
         text_label_top = text_label_top - text_gap
         -- Labels to show "Trained by: <trainers> or Sold by: <vendors> or Dropped by: <mobs> or Obtained from: <quest>"
-        self.labels.alt_sources.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":", text_label_left, text_label_top, "LABEL", "TOPLEFT")
+        self.labels.alt_sources.title = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["also learned from"][MTSLUI_CURRENT_LANGUAGE], text_label_left, text_label_top, "LABEL", "TOPLEFT")
         -- Create a frame to show the alternative sources
         self.labels.alt_sources.ui_frame = MTSLUI_TOOLS:CreateBaseFrame("Frame", "", self.ui_frame, nil, self.FRAME_WIDTH - 118, self.FRAME_ALT_SOURCES_HEIGHT, false)
         -- position halfway from the frame with primary sources
@@ -149,7 +149,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.alt_sources.ui_frame:Show()
         self.labels.alt_sources.values = {}
         text_label_top = -5
-        text_label_right = 10
+        text_label_right = 15
         -- Create the labels for sources
         for i=1,self.MAX_SOURCES_SHOWN_SECONDARY do
             local string_sources_content = MTSLUI_TOOLS:CreateLabel(self.labels.alt_sources.ui_frame, i .. " alt", text_label_right, text_label_top, "TEXT", "TOPLEFT")
@@ -242,14 +242,14 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.special_action.value:SetText("-")
         self.labels.price.value:SetText("-")
         self.labels.type.value:SetText("-")
-        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE])
         self.labels.sources.title:Hide()
         self.labels.sources.ui_frame:Hide()
-        self.labels.alt_type.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["also learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+        self.labels.alt_type.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["also learned from"][MTSLUI_CURRENT_LANGUAGE])
         self.labels.alt_type.title:Hide()
         self.labels.alt_type.value:SetText("-")
         self.labels.alt_type.value:Hide()
-        self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+        self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["learned from"][MTSLUI_CURRENT_LANGUAGE])
         self.labels.alt_sources.title:Hide()
         self.labels.alt_sources.ui_frame:Hide()
         -- hide the tooltips
@@ -395,7 +395,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
     SetRequiredSpecialAction = function(self, special_action)
         -- if special action is required
         if special_action ~= nil then
-            self.labels.special_action.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. special_action["English"])
+            self.labels.special_action.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. special_action[MTSLUI_CURRENT_LANGUAGE])
         else
             self.labels.special_action.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. "-")
         end
@@ -410,7 +410,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         self.labels.price.value:SetText(MTSL_TOOLS:GetNumberAsMoneyString(trainers_info.price))
         self.labels.type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["trainer"][MTSLUI_CURRENT_LANGUAGE])
         -- Labels to show "Trained by: <trainer>"
-        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["trained by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["trained by"][MTSLUI_CURRENT_LANGUAGE])
         -- Get all "available" trainers for the player
         local trainers = MTSL_LOGIC_PLAYER_NPC:GetNpcsByIds(trainers_info.sources)
         -- Show all the npcs
@@ -437,16 +437,16 @@ MTSLUI_SKILL_DETAIL_FRAME = {
 
         local quest = MTSL_LOGIC_QUEST:GetQuestByIds(quest_ids)
         if is_alternative_source == 1 then
-            self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["started by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+            self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["started by"][MTSLUI_CURRENT_LANGUAGE])
         else
-            self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["started by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+            self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["started by"][MTSLUI_CURRENT_LANGUAGE])
         end
         -- check if quest is availbe to us
         if quest ~= nil then
             if is_alternative_source == 1 then
-                self.labels.alt_type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["quest"][MTSLUI_CURRENT_LANGUAGE] .. ": " .. quest["name"][MTSLUI_CURRENT_LANGUAGE])
+                self.labels.alt_type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["quest"][MTSLUI_CURRENT_LANGUAGE] .. " : " .. quest["name"][MTSLUI_CURRENT_LANGUAGE])
             else
-                self.labels.type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["quest"][MTSLUI_CURRENT_LANGUAGE] .. ": " .. quest["name"][MTSLUI_CURRENT_LANGUAGE])
+                self.labels.type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["quest"][MTSLUI_CURRENT_LANGUAGE] .. " : " .. quest["name"][MTSLUI_CURRENT_LANGUAGE])
             end
             -- show xp level requirements if any
             self:SetRequiredXPLevel(quest.min_xp_level, current_xp_level)
@@ -552,7 +552,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
                 -- obtained from vendors
                 if has_vendors > 0 then
                     self.tooltip_source_name = "item:" .. item_id
-                    self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["sold by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+                    self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["sold by"][MTSLUI_CURRENT_LANGUAGE])
                     self.labels.price.value:SetText(MTSL_TOOLS:GetNumberAsMoneyString(item.vendors.price))
                     -- Get all "available" vendors for the player
                     local vendors = MTSL_LOGIC_PLAYER_NPC:GetNpcsByIds(item.vendors.sources)
@@ -573,7 +573,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
                 if has_drops > 0 then
                     -- primary source since no vendors or quests
                     if has_vendors <= 0 and has_quests <= 0 then
-                        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["dropped by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+                        self.labels.sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["dropped by"][MTSLUI_CURRENT_LANGUAGE])
                         -- check if drop of mob or world drops
                         if item.drops.mobs_range ~= nil then
                             self:ShowWorldDropSources(item.drops.mobs_range.min_xp_level, item.drops.mobs_range.max_xp_level, 0)
@@ -585,7 +585,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
                     -- alternative/secundary source
                     else
                         self.labels.alt_type.value:SetText(MTSLUI_FONTS.COLORS.TEXT.NORMAL .. MTSLUI_LOCALES_LABELS["mobs"][MTSLUI_CURRENT_LANGUAGE])
-                        self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["dropped by"][MTSLUI_CURRENT_LANGUAGE] .. ":")
+                        self.labels.alt_sources.title:SetText(MTSLUI_FONTS.COLORS.TEXT.TITLE .. MTSLUI_LOCALES_LABELS["dropped by"][MTSLUI_CURRENT_LANGUAGE])
                         -- check if drop of mob or world drops
                         if item.drops.mobs_range ~= nil then
                             self:ShowWorldDropSources(item.drops.mobs_range.min_xp_level, item.drops.mobs_range.max_xp_level, 1)
@@ -651,12 +651,11 @@ MTSLUI_SKILL_DETAIL_FRAME = {
                     end
                     -- Check if require reputation to interact with npc
                     if npcs[i].reputation ~= nil then
-                        text = text .. " [Requires ".. npcs[i].reputation.faction .. " - " .. npcs[i].reputation.level .. "]"
+                        text = text .. " [".. npcs[i].reputation.faction .. " - " .. npcs[i].reputation.level .. "]"
                     end
                     -- Check if special action is required to interact with npc
                     if npcs[i].special_action ~= nil then
-                        -- TODO trasnlate all special actions
-                        text = text .. " [ ".. npcs[i].special_action["English"] .. "]"
+                        text = text .. " [".. npcs[i].special_action[MTSLUI_CURRENT_LANGUAGE] .. "]"
                     end
                     -- show level of mob/npc if known
                     if npcs[i].level ~= nil and npcs[i].level ~= "" and npcs[i].level ~= 0 then
@@ -745,7 +744,7 @@ MTSLUI_SKILL_DETAIL_FRAME = {
         labels_sources.ui_frame:Show()
         for i=1, amount_labels do
             if objects[i] ~= nil then
-                local text =  "[object] " .. objects[i]["name"][MTSLUI_CURRENT_LANGUAGE] .. " - " ..  MTSL_LOGIC_WORLD:GetZoneNameById(objects[i].zone_id)
+                local text =  "[" .. MTSLUI_LOCALES_LABELS["object"][MTSLUI_CURRENT_LANGUAGE] .. "]" .. objects[i]["name"][MTSLUI_CURRENT_LANGUAGE] .. " - " ..  MTSL_LOGIC_WORLD:GetZoneNameById(objects[i].zone_id)
                 -- add coords if known
                 if objects[i].location ~= nil and objects[i].location.x ~= "-" and
                         objects[i].location.x ~= "" then