+++ Interface/AddOns/MissingTradeSkillsList/UI/Options/ResetFrame.lua
@@ -23,14 +23,18 @@ MTSLOPTUI_RESET_FRAME = {
         -- create list of realms & list with players on a realm
         self:BuildRealmList()
         self:BuildPlayersOnCurrentRealmList()
+
+        self.ui_frame.realm_text = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["realm"][MTSLUI_CURRENT_LANGUAGE], 22, -30, "LABEL", "TOPLEFT")
         -- add drop down list for realm & characters
         self.ui_frame.realm_drop_down = CreateFrame("Frame", "MTSLOPTUI_RESET_FRAME_DD_REALM", self.ui_frame, "UIDropDownMenuTemplate")
-        self.ui_frame.realm_drop_down:SetPoint("TOPLEFT", self.ui_frame, "TOPLEFT", 5, -26)
+        self.ui_frame.realm_drop_down:SetPoint("TOPLEFT", self.ui_frame, "TOPLEFT", 70, -22)
         self.ui_frame.realm_drop_down.initialize = self.CreateDropDownRealms
-        UIDropDownMenu_SetWidth(self.ui_frame.realm_drop_down, 200)
+        UIDropDownMenu_SetWidth(self.ui_frame.realm_drop_down, 150)
         UIDropDownMenu_SetText(self.ui_frame.realm_drop_down, self.current_realm)
+
+        self.ui_frame.player_text = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_LOCALES_LABELS["character"][MTSLUI_CURRENT_LANGUAGE], 280, -30, "LABEL", "TOPLEFT")
         self.ui_frame.player_drop_down = CreateFrame("Frame", "MTSLOPTUI_RESET_FRAME_DD_PLAYER", self.ui_frame, "UIDropDownMenuTemplate")
-        self.ui_frame.player_drop_down:SetPoint("TOPLEFT", self.ui_frame.realm_drop_down, "TOPRIGHT", -25, 0)
+        self.ui_frame.player_drop_down:SetPoint("TOPLEFT", self.ui_frame.realm_drop_down, "TOPRIGHT", 65, 0)
         self.ui_frame.player_drop_down.initialize = self.CreateDropDownPlayersOnRealm
         UIDropDownMenu_SetWidth(self.ui_frame.player_drop_down, 150)
         UIDropDownMenu_SetText(self.ui_frame.player_drop_down, self.current_player)
@@ -39,8 +43,8 @@ MTSLOPTUI_RESET_FRAME = {
         local left = MTSLUI_OPTIONS_MENU_FRAME.FRAME_WIDTH - 230
         -- Remove character button
         self.remove_btn = MTSLUI_TOOLS:CreateBaseFrame("Button", "MTSLOPTUI_RemoveChar_Button", self.ui_frame, "UIPanelButtonTemplate", self.BUTTON_WIDTH - 50, self.BUTTON_HEIGHT)
-        self.remove_btn:SetPoint("TOPLEFT", self.ui_frame, "TOPLEFT", left + 50, -25)
-        self.remove_btn:SetText("Remove character") -- MTSLUI_LOCALES_LABELS["save"][MTSLUI_CURRENT_LANGUAGE])
+        self.remove_btn:SetPoint("TOPLEFT", self.ui_frame, "TOPLEFT", left + 50, -21)
+        self.remove_btn:SetText(MTSLUI_LOCALES_LABELS["delete"][MTSLUI_CURRENT_LANGUAGE])
         self.remove_btn:SetScript("OnClick", function ()
             -- only remove if realm & player chosen
             if MTSLOPTUI_RESET_FRAME.current_realm ~= nil and MTSLOPTUI_RESET_FRAME.current_player ~= nil then
@@ -65,10 +69,10 @@ MTSLOPTUI_RESET_FRAME = {
             end
         end)
         -- Reset all
-        self.ui_frame.reset_text = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_FONTS.COLORS.TEXT.ERROR .. "WARNING: can not be undone!", 25, -75, "TITLE", "TOPLEFT")
+        self.ui_frame.reset_text = MTSLUI_TOOLS:CreateLabel(self.ui_frame, MTSLUI_FONTS.COLORS.TEXT.ERROR .. MTSLUI_LOCALES_LABELS["permanent"][MTSLUI_CURRENT_LANGUAGE], 25, -75, "TITLE", "TOPLEFT")
         self.reset_btn = MTSLUI_TOOLS:CreateBaseFrame("Button", "MTSLOPTUI_ResetAll_Button", self.ui_frame, "UIPanelButtonTemplate",  self.BUTTON_WIDTH + 50, self.BUTTON_HEIGHT)
         self.reset_btn:SetPoint("TOPLEFT", self.ui_frame, "TOPLEFT", left - 50, -65)
-        self.reset_btn:SetText("Remove all saved characters") --MTSLUI_LOCALES_LABELS["cancel"][MTSLUI_CURRENT_LANGUAGE])
+        self.reset_btn:SetText(MTSLUI_LOCALES_LABELS["delete all"][MTSLUI_CURRENT_LANGUAGE])
         self.reset_btn:SetScript("OnClick", function ()
             MTSL_LOGIC_SAVED_VARIABLES:RemoveAllCharacters()
         end)